+++ agora/agora/spec.py
@@ -34,9 +34,6 @@ CUSTOM_USERS = {
         'date_joined': {
             '.flag.nowrite': {},
             '.field.datetime': {}},
-        'avatar': {
-            '.field.file': {},
-            'default': ''},
         'shibboleth_id': {
             '.field.string': {},
             '.flag.nowrite': {},
@@ -195,6 +192,11 @@ MERIL_DOMAINS = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -217,6 +219,11 @@ DOMAINS = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -249,6 +256,11 @@ MERIL_SUBDOMAINS = {
             '.field.string': {},
             '.flag.searchable': {},
             '.flag.nullable.default': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -277,6 +289,11 @@ SUBDOMAINS = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -299,6 +316,11 @@ SUPERCATEGORIES = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -327,6 +349,11 @@ CATEGORIES = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -355,6 +382,11 @@ SUBCATEGORIES = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -377,6 +409,11 @@ LEGAL_STATUSES = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -403,6 +440,11 @@ NETWORKS = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -429,6 +471,11 @@ STRUCTURES = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -477,6 +524,11 @@ ESFRI_DOMAINS= {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -499,6 +551,11 @@ ESFRI_TYPES= {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -521,6 +578,11 @@ ACTIVITIES= {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -543,6 +605,11 @@ FUNDING_BODIES= {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -565,6 +632,11 @@ FUNDING_PROGRAMS = {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -590,6 +662,11 @@ ORDER_TYPES= {
         'description': {
             '.field.string': {},
             '.flag.nullable.default': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -616,6 +693,11 @@ ACCESS_TYPES= {
             '.field.string': {},
             '.flag.searchable': {},
             '.flag.nullable.default': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -642,6 +724,11 @@ ACCESS_MODES= {
             '.field.string': {},
             '.flag.searchable': {},
             '.flag.nullable.default': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -668,6 +755,11 @@ TRLS= {
             '.field.string': {},
             '.flag.searchable': {},
             '.flag.nullable.default': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -694,6 +786,11 @@ RESOURCE_LIFECYCLE_STATUSES= {
             '.field.string': {},
             '.flag.searchable': {},
             '.flag.nullable.default': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -716,6 +813,11 @@ CHALLENGES= {
             '.field.string': {},
             '.flag.orderable': {},
             '.flag.searchable': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -737,6 +839,11 @@ ORGANISATIONS = {
             '.field.uuid': {},
             '.flag.nowrite': {}},
             '.flag.orderable': {},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
         'created_at': {
             '.flag.nowrite': {},
             '.field.datetime': {}},
@@ -1122,6 +1229,11 @@ TARGET_USERS = {
             '.field.string': {},
             '.flag.searchable': {},
             '.flag.nullable.default': {}},
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -1650,6 +1762,18 @@ RESOURCES = {
             'default': 'draft',
             '.flag.filterable': {},
         },
+        'eosc_id': {
+            '.field.string': {},
+            '.flag.nullable.default': {},
+            '.flag.searchable': {},
+            '.flag.orderable': {}},
+        'eosc_published_at': {
+            '.flag.nowrite': {},
+            '.field.datetime': {}},
+        'eosc_state': {
+            '.field.string': {},
+            'default': 'draft',
+            '.flag.filterable': {}},
     },
     'actions': {
         '.action-template.django.list': {},
@@ -1672,6 +1796,20 @@ RESOURCES = {
                     )},
                 },
             },
+        },
+        'publish-eosc': {
+            '.action.django.recipe.partial_update': {},
+            'method': 'POST',
+            'status_code': 200,
+            'url': '/*/publish-eosc/',
+            ':custom_update_handler': 'agora.actions.resource_publish_eosc',
+        },
+        'update-eosc': {
+            '.action.django.recipe.partial_update': {},
+            'method': 'POST',
+            'status_code': 200,
+            'url': '/*/update-eosc/',
+            ':custom_update_handler': 'agora.actions.resource_update_eosc',
         },
     },
 }