+++ ui/app/locales/en/privacy_policy.js
@@ -1,0 +1,7 @@
+const privacy_policy = {
+  'page': {
+    'title': 'Privacy Policy',
+  },
+};
+
+export { privacy_policy };