+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0](https://github.com/grnet/agora-sp/compare/v1.1.3...v1.2.0) - 2021-03-29
+
+### Added
+- Add Terms of Use page (disabled by default).
+- Add Access Policy page (disabled by default).
+- Publish resource action to EOSC portal.
+- Expose eosc_id to various resources.
+
+### Changed
+- Move Privacy Policy to dedicated url.
+- Resource's changelog, open source, standards and certification fields are now 
+textareas in the UI.
+
 ## [1.1.3](https://github.com/grnet/agora-sp/compare/v1.1.2...v1.1.3) - 2021-02-03
 
 ### Added
@@ -19,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Expose epp_loi_5_country_or_territory field as capitalized instead of lowercase.
+
 
 ## [1.1.1](https://github.com/grnet/agora-sp/compare/v1.1.0...v1.1.1) - 2020-12-11